# backend/api/openapi.yaml
openapi: 3.0.0
info:
  title: Tapiro API
  description: API for AI-Driven Personalized Advertising System
  version: 1.0.0
  contact: { name: Tapiro Support, email: tapirosupport@gmail.com }
...
servers:
  - url: https://virtserver.swaggerhub.com/CHAMATHDEWMINA25/TAPIRO/1.0.0
    description: SwaggerHub API Auto Mocking
  - url: https://api.tapiro.com/v1
    description: Production API
  - url: http://localhost:3000
    description: Development API
security:
  - oauth2:
      - user:read
      - user:write
      - store:read
      - store:write
paths:
...
  /stores/{storeId}/ads:
    post:
      tags: [Ad Management]
      summary: Create new advertisement
      description: Create a new targeted advertisement for a specific store. Allows setting categories, media, and validity period.
      operationId: createAd
      parameters:
        - name: storeId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Ad"
        required: true
      responses:
        "201":
          description: Advertisement successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
      security:
        - oauth2: [store:write]
      x-swagger-router-controller: Advertisements

    get:
      tags: [Ad Management]
      summary: List store ads
      description: Retrieve all active advertisements for a specific store including their targeting criteria and status
      operationId: listAds
      parameters:
        - name: storeId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Successful retrieval of store advertisements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
      security:
        - oauth2: [store:read]
      x-swagger-router-controller: Advertisements
  /stores/{storeId}/ads/{adId}:
    put:
      tags: [Ad Management]
      summary: Update specific ad
      description: Modify an existing advertisement's details including targeting parameters and validity period
      operationId: updateAd
      parameters:
        - name: storeId
          in: path
          required: true
          schema: { type: string }
        - name: adId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Ad"
        required: true
      responses:
        "200":
          description: Advertisement successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
      security:
        - oauth2: [store:write]
      x-swagger-router-controller: Advertisements

    delete:
      tags: [Ad Management]
      summary: Delete specific ad
      description: Permanently remove an advertisement from the store's active campaigns
      operationId: deleteAd
      parameters:
        - name: storeId
          in: path
          required: true
          schema: { type: string }
        - name: adId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Advertisement successfully deleted
      security:
        - oauth2: [store:write]
      x-swagger-router-controller: Advertisements
...
components:
  schemas:
...
    Ad:
      type: object
      required: [title, target_categories]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title: { type: string }
        description: { type: string }
        target_categories:
          type: array
          items: { type: string }
        media_url: { type: string }
        validity_period:
          type: object
          properties:
            start: { type: string, format: date-time }
            end: { type: string, format: date-time }
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://dev-zuxebycdcmuazvo8.us.auth0.com/authorize
          tokenUrl: https://dev-zuxebycdcmuazvo8.us.auth0.com/oauth/token
          scopes:
            user:read: Read user data
            user:write: Modify user data
            store:read: Read store data
            store:write: Modify store data
...